+++ appnea/lib/routers/routers.dart
@@ -1,0 +1,20 @@
+import 'package:appnea/screens/home.dart';
+import 'package:appnea/screens/m_menu.dart';
+import 'package:go_router/go_router.dart';
+
+// GoRouter configuration
+final router = GoRouter(
+  initialLocation: '/',
+  routes: [
+    GoRoute(
+      name: 'home', // Optional, add name to your routes. Allows you navigate by name instead of path
+      path: '/',
+      builder: (context, state) => const MyHomePage(title: "Titulo"),
+    ),
+    GoRoute(
+      name: 'home2',
+      path: '/home',
+      builder: (context, state) => const HomePage(),
+    ),
+  ],
+);
